pyproject.toml
[tool.poetry]
name = "lblcs_bot"
version = "0.1.0"
description = "Scrapes information from servers such as roles, users and inserts it into a database for the LBLCS."
authors = ["Alexander Summer"]
license = "MIT"
readme = "README.md"
packages=[{include = "lblcs_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.2"
python-dotenv = "^1.0.0"
cachetools = "^5.3.2"
supabase = "^2.3.4"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
